requirements.txt
langchain
langgraph
faiss-cpu
fastapi
uvicorn
python-dotenv
pypdf
langchain-groq
pypdf
langchain-core
langchain-community
langchain-text-splitters
faiss-cpu 
sentence-transformers
streamlit
requests
psycopg2-binary
langchain-huggingface